requirements.txt
astroid==2.8.0
attrs==21.2.0
autopep8==1.5.7
black==21.9b0
click==8.0.1
coverage==6.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
flask-swagger-ui==3.36.0
greenlet==1.1.1
gunicorn==20.1.0
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
linecache2==1.0.0
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
passlib==1.7.4
pathspec==0.9.0
pep8==1.7.1
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.11.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
PyYAML==5.4.1
regex==2021.9.30
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
tomli==1.2.1
traceback2==1.4.0
typing-extensions==3.10.0.2
unittest2==1.1.0
Werkzeug==2.0.1
wrapt==1.12.1
